requirements.txt
beautifulsoup4==4.8.2
biopython==1.79
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
coloredlogs==7.3
cwltool==2.0.20200224214940
decorator==4.4.2
html5lib==1.0.1
humanfriendly==4.18
idna==2.8
isodate==0.6.0
joblib==1.1.0
lockfile==0.12.2
lxml==4.5.0
mistune==0.8.4
msgpack==0.6.2
mypy-extensions==0.4.3
networkx==2.4
numpy==1.22.2
pandas==1.4.1
Pillow==9.1.1
prov==1.5.2
psutil==5.5.1
pyparsing==2.4.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.3.1
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.22.0
ruamel.yaml==0.15.89
schema-salad==5.0.20200220195218
scikit-learn==1.0.2
scipy==1.8.0
semver==2.13.0
shellescape==3.4.1
six==1.14.0
soupsieve==1.9.5
SPARQLWrapper==1.8.5
spython==0.1.18
threadpoolctl==3.1.0
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchvision==0.11.2+cu113
typing-extensions==3.7.4.1
urllib3==1.25.8
webencodings==0.5.1
